Subject: Modify default build behavior
Description:

Update the build to including MPI and OFI by default as well as to
try including the SS11 libs from the host by default if they are found.
The intent of this is to make it easier for users to run the containers
without needing to worry about SS libraries or bind-mounts.