Subject: Update OFI/OMPI/AWS components and cleanup build
Description:

This mod updates several of the key components, including:
  - libfabric 1.x -> 2.0
  - OMPI 4.1.x -> 5.x
  - AWS 1.9.2 -> 1.13.1

A key part of this work is to pull in the Cray source for CXI
and libfabric and compile them inside the container. By building from
source inside the container we can ensure the libraries link against
the right version of libraries, to prevent issues with mismatches if we
bind-mount in the libraries from the host system. This also makes it
easier to run on the host system since no libraries should need to
be mounted from the host.

With the addition of the Cray CXI/libfabric, the installation directory
for the HPC libs has been changed to be under /container/hpc. This root
directory is now used for all HPC related tools compiled and installed
into the container (e.g, CXI, OFI, OMPI and AWS).

Also in this mod are several cleanup bits to the build scripts to remove
commented out code and simplify things. Further, the scrape_libs entrypoint
was modified to provide default settings for important NCCL/OFI variables
and to try working around a potential race condition related to mismatches
in cuda driver between the host and container.

Finally, a new script was added that sets up some env variables globally
inside the container, apart from the entrypoint script, so that the
settings help at container build time as well as runtime. These settings
are intended to help gcc/ld during container build time to find the
right HPC related headers/libraries.