Subject: Fix missing curl package and change default MPI
Description:

This mod adds a possibly missing curl package that is needed by
libfabric for the curl.h, etc. This package may not exist, depending
on the base docker image, so install it to make sure libfabric
can use it.

Also, if we install MPI that targets libfabric, replace the OMPI
installed in /usr/local/mpi, which could be there from the base
image but would likely target IB.