Subject: Make horovod build not happen if pip does not exist
Description:

This mod disables the install of horovod if pip does not exist in the base
image. This is possible for some base images from NGC that do not include
PyTorch, such as the HPC image for running the HPL benchmarks.

Note that this could also be done by not having the build_horovod.sh
exist in the Dockerfile-ngc-hpc.